1st test of clsAccount(tstAccount.cs created)
